Remove unnecessary cast
getContext() returns short so casting to short is not necessary.

+++ gerrit-gwtui/src/main/java/com/google/gerrit/client/patches/PatchScriptSettingsPanel.java
@@ -314,7 +314,7 @@ public class PatchScriptSettingsPanel extends Composite {
     if (0 <= sel) {
       return Short.parseShort(context.getValue(sel));
     }
-    return (short) getValue().getContext();
+    return getValue().getContext();
   }
 
   private void setContext(int ctx) {
